Switch to blackbox testing so rocksdb is not rebuilt when using "go test".
The "go test" command will rebuild a package being tested every time. This is
normal behavior.
https://golang.org/cmd/go/#hdr-Test_packages

Rebuilding rocksdb takes a substantial amount of time. We could avoid this
rebuild by switching to blackbox testing with "package gorocksdb_test" in our
*_test.go files. None of the tests depend on anything unexported so this change
is not hard to make.

To observe the difference "go install" the package and then "go test". You will
notice that repeated "go test" invocations are not causing the gorocksdb
package to be rebuilt.